pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vgr_sdk"
version = "0.1.0"
description = "VGR SDK: Robot-side subscriber & task manager for vision-guided manipulation"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
  {name = "Your Team", email = "team@example.com"}
]
keywords = ["robotics", "vision", "manipulation", "VGR", "ROS"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Topic :: Scientific/Engineering",
  "Topic :: Software Development :: Libraries",
]

dependencies = [
  "numpy>=1.24",
  "PyYAML>=6.0",
  "jsonschema>=4.21",
  "typing-extensions>=4.10",
  "msgpack>=1.0",         # handy for future binary/fast transport
  "simplejson>=3.19",     # tolerant JSON (optional, but useful)
]

[project.optional-dependencies]
dev = [
  "mypy>=1.9",
  "ruff>=0.4",
  "pytest>=8.0",
  "pytest-cov>=5.0",
]
# NOTE: ROS Python deps (rospy, sensor_msgs, etc.) come from your ROS distro (apt),
# not from pip; keep them out of dependencies.

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["tests", "scripts"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E","F","I","B","UP","N"]
ignore = []
